backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
pyjwt==2.8.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# OAuth2 & OIDC
authlib==1.2.1
httpx==0.25.2

# AWS
boto3==1.29.7
pycognito==2023.5.0

# Configuration & Validation
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# CORS & Middleware
python-cors==1.0.0

# Rate Limiting
slowapi==0.1.9

# Utilities
requests==2.31.0
